requirements.txt
agate==1.6.3
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.8
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
Arpeggio==2.0.2
arrow==1.3.0
asgiref==3.8.1
astropy==6.1.1
astropy-iers-data==0.2024.7.15.0.31.42
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
black==24.10.0
bleach==6.1.0
blinker==1.8.2
Brlapi==0.8.3
Brotli==1.0.9
business-rules==1.1.1
cachetools==5.4.0
certifi==2021.10.8
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.6
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
csvkit==1.0.6
cupshelpers==1.0
cycler==0.11.0
dbfread==2.0.7
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
django-cors-headers==4.4.0
dnspython==2.1.0
duplicity==0.8.21
entrypoints==0.4
et-xmlfile==1.0.1
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-GraphQL==2.0.0
fonttools==4.29.1
fqdn==1.5.1
frozendict==1.2
fs==2.4.12
future==0.18.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.43
gpg==1.16.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
graphql-server-core==1.1.1
greenlet==1.1.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.2
idp-engine==0.11.1
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.24.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jdcal==1.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsons==1.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
leather==0.3.4
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mock==5.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.2.0
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openai==1.55.2
openpyxl==3.0.9
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==2.9.3
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==5.27.2
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyarrow==17.0.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.2
pydantic_core==2.27.1
pyDatalog==0.17.4
pydeck==0.9.1
pyerfa==2.0.1.4
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pyknow @ file:///home/abishek/pyknow
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==8.2.2
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-drools-sdk==0.0.3
python-json-logger==2.0.7
python-magic==0.4.24
python-slugify==4.0.0
pythran==0.10.0
pytimeparse==1.1.5
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
schema==0.6.7
scikit-learn==1.5.0
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.26.1
Send2Trash==1.8.3
six==1.16.0
sklearn-preprocessing==0.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.31
sqlparse==0.4.2
stack-data==0.6.3
streamlit==1.36.0
sympy==1.9
systemd-python==234
tenacity==8.5.0
terminado==0.18.1
testpath==0.5.0
textX==2.3.0
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torbrowser-launcher==0.3.3
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
typish==1.9.3
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
Unidecode==1.3.3
uri-template==1.3.0
urllib3==1.26.7
usb-creator==0.3.7
wadllib==1.3.6
watchdog==4.0.1
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.2
widgetsnbextension==2.0.0
wifite==2.2.5
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
z3-solver==4.12.4.0
zipp==1.0.0
